CDRIVER-4231: New random generation of OID values (#931)
* New random generation of OID values [Fixes CDRIVER-4231]

* Restore hashing of machine attributes, plus an atomic init-counter,
  plus the process ID.
* Instead of MD5 hashing, use SipHash.
* No more magic numbers, make bson-context -Wconversion clean, comment what's happening in the random init.
* Remove old deprecated docs and no-op tests.
* Host name mocking is gone

+++ src/libbson/src/bson/bson-endian.h
@@ -133,7 +133,7 @@ BSON_BEGIN_DECLS
 static BSON_INLINE uint16_t
 __bson_uint16_swap_slow (uint16_t v) /* IN */
 {
-   return ((v & 0x00FF) << 8) | ((v & 0xFF00) >> 8);
+   return (uint16_t) ((v & 0x00FF) << 8) | (uint16_t) ((v & 0xFF00) >> 8);
 }
